46-添加升级时候对金钱的消耗

+++ Assets/Scripts/Enemy.cs
@@ -6,8 +6,8 @@ using UnityEngine.UI;
 public class Enemy : MonoBehaviour
 {
     public float speed = 10;
-    public int hp = 150;
-    private int totalHp;
+    public float hp = 150;
+    private float totalHp;
     public GameObject explosionEffect;
     public int money = 300;
     public Slider hpSlider;
@@ -54,7 +54,7 @@ public class Enemy : MonoBehaviour
         EnemySpawner.CountEnemyAlive--;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if (hp <= 0) return;
         hp -= damage;
